+++ lib/mistral_sgela_ai.dart
@@ -2,7 +2,6 @@
 ///
 library;
 
-export 'src/dart_mistral_api_base.dart';
 export 'src/data/mistral.dart';
 export 'src/service/dio_util.dart';
 export 'src/service/mistral_service.dart';