+++ CHANGELOG.md
@@ -1,18 +1,56 @@
 ## 1.0.0
 
-- Initial version of dart_mistral.
-
-## 1.0.0+1
-
-- Cleaned up 
+- Initial version of mistral_sgela_ai. This version is the OG base!
 
 
 ## 1.0.1
 
-- Cleaned up continued
+- Refactored something
+
 
 ## 1.0.2
 
+- Updated README
+
+## 1.0.3
+
+- Updated list of classes
+
+
+## 1.0.4
+
+- Updated constants for model names
+
+## 1.0.5
+
+- Updated constants for model names, again
+
+## 1.0.6
+
+- Added hello (ping) method
+
+
+## 1.0.7
+
 - Refactored README
 
+## 1.0.8
+
+- Refactored README - added hello text
+
+## 1.0.9
+
+- Updated example, links in README
+
+## 1.0.10
+
+- Refactored Mistral request
+
+## 1.0.11
+
+- Updated debug logging
+
+## 1.1.0
+
+- Refactored Mistral request error handling